% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_proj_db.R
\name{carbon_proj_db_to_file}
\alias{carbon_proj_db_to_file}
\title{Export Carbon Project Database to Parquet File}
\usage{
carbon_proj_db_to_file(
  dest_path = "forest_carbon_boundaries.parquet",
  db_path = NULL
)
}
\arguments{
\item{dest_path}{The destination path for the output Parquet file. If not provided, it defaults to "forest_carbon_boundaries.parquet" in the current working directory.}

\item{db_path}{The path to the carbon project database. If not provided, it will attempt to find the database path using the \code{wheredd_db_path()} function.}
}
\value{
The path to the exported Parquet file, invisibly.
}
\description{
Exports the carbon project database to a Parquet file. The function connects to the database, retrieves the data, and writes it to a specified Parquet file.
}
